Imported clothing
Fixed boot's normals since you could see through some faces at the bottom of the shoe. After, imported all of the modular parts as skeletal meshes and the non-modular parts are static meshes. Added blueprints to construction blueprint to apply the skeletal meshes to the player mesh and have the animations work in tandem.